+++ lib/Controller.php
@@ -30,7 +30,7 @@ class Controller
      *
      * @const string
      */
-    const VERSION = '2.0.0';
+    const VERSION = '2.0.1';
 
     /**
      * minimal required PHP version