+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0.1 (not yet released)
 * CHANGED: CSP recommendation around bootstrap5 template resolved in Firefox 131 (#1613)
 * FIXED: Allow pasting a password for decrypting a paste (#1620)
+* FIXED: Allow copying the shortened link after using a URL shortener (#1624)
 
 ## 2.0.0 (2025-07-28)
 * ADDED: Error logging in database and filesystem backend (#1554)