+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # PrivateBin version history
 
 ## 2.0.1 (not yet released)
+* ADDED: Auto shorten URLs with config option `shortenbydefault` (#1627)
+* ADDED: Added `shortenviashlink` endpoint with an `shlink` configuration section
+* ADDED: Password peek (#1254)
 * CHANGED: CSP recommendation around bootstrap5 template resolved in Firefox 131 (#1613)
 * FIXED: Allow pasting a password for decrypting a paste (#1620)
 * FIXED: Allow copying the shortened link after using a URL shortener (#1624)
-* ADDED: Auto shorten URLs with config option `shortenbydefault` (#1627)
-* ADDED: Added `shortenviashlink` endpoint with an `shlink` configuration section 
-* FIXED: Check for quotes and conical braces when extracting short url (#1644)
+* FIXED: URL extraction fails when frame-ancestors is set in CSP (#1644)
 
 ## 2.0.0 (2025-07-28)
 * ADDED: Error logging in database and filesystem backend (#1554)